Fix bug where loading new level lagged

+++ GameLib/Game.h
@@ -93,7 +93,7 @@ private:
     bool mLevelWon = false;
 
     /// Records time of level victory
-    long mVictoryTime = -1;
+    int mVictoryTime = -1;
 
 public:
     Game();
